+++ R/table.R
@@ -11,7 +11,7 @@ NULL
 #'
 #' @examples
 #' \dontrun{
-#' conn <- Td(apikey="xxxxx")
+#' conn <- Td(apikey = "xxxxx")
 #' exist_table(conn, "mydb", "iris")
 #' }
 #'
@@ -30,7 +30,7 @@ exist_table <- function(conn, dbname, table) {
 #'
 #' @examples
 #' \dontrun{
-#' conn <- Td(apikey="xxxxx")
+#' conn <- Td(apikey = "xxxxx")
 #' list_tables(conn, "mydb")
 #' }
 #'
@@ -38,7 +38,7 @@ exist_table <- function(conn, dbname, table) {
 #'
 list_tables <- function(conn, dbname) {
   res <- .get(conn, paste0("/v3/table/list/", dbname))
-  return(as.data.frame(do.call("rbind", res$tables)))
+  return(dplyr::tibble(as.data.frame(do.call("rbind", res$tables))))
 }
 
 
@@ -51,7 +51,7 @@ list_tables <- function(conn, dbname) {
 #'
 #' @examples
 #' \dontrun{
-#' conn <- Td(apikey="xxxx")
+#' conn <- Td(apikey = "xxxx")
 #' create_table(conn, "mydb", "new_table")
 #' }
 #'
@@ -71,7 +71,7 @@ create_table <- function(conn, dbname, table) {
 #'
 #' @examples
 #' \dontrun{
-#' conn <- Td(apikey="xxxxx")
+#' conn <- Td(apikey = "xxxxx")
 #' delete_table(conn, "mydb", "iris")
 #' }
 #'
@@ -79,5 +79,32 @@ create_table <- function(conn, dbname, table) {
 #'
 delete_table <- function(conn, dbname, table) {
   res <- .post(conn, paste0("/v3/table/delete/", dbname, "/", table), character(0))
-  return(res$type)
+  return(TRUE)
+}
+
+#' Update schema of a table
+#'
+#' @param conn \code{Td} connection
+#' @param dbname Data base name
+#' @param table Table name
+#' @param schema Schema of the table to be updated
+#' @return Returns \code{TRUE} or \code{FALSE}, whether the execution succeeded or not.
+#'
+#' @examples
+#' \dontrun{
+#' conn <- Td(apikey = "xxxxx")
+#' s <- rbind(
+#' c("sepal_length", "double", "sepal_length"),
+#' c("sepal_width", "double", "sepal_width"),
+#' c("petal_length", "double", "petal_length"),
+#' c("petal_width", "double", "petal_width"),
+#' c("species", "string", "species"))
+#' udpate_schema(conn, "mydb", "iris", s)
+#' }
+#'
+#' @export
+#'
+update_schema <- function(conn, dbname, table, schema) {
+  res <- .post(conn, paste0("/v3/table/update-schema/", dbname, "/", table), query=paste0("schema=", jsonlite::toJSON(schema)))
+  return(TRUE)
 }