+++ tests/testthat/test-td.R
@@ -1,13 +1,13 @@
-context('td')
+context("td")
 
 library(mockery)
 
-con <- Td(apikey="xxxxx")
-embulk_exec <- if(.Platform$OS.type == "windows") "embulk.bat" else "embulk"
+con <- Td(apikey = "xxxxx")
+embulk_exec <- if (.Platform$OS.type == "windows") "embulk.bat" else "embulk"
 
-
-test_that("td_upload works with mock",{
-  template_path <- system.file("extdata", "tsv_upload.yml.liquid", package="RTD")
+# TODO: test for "bulk_import" mode
+test_that("td_upload works with mock", {
+  template_path <- system.file("extdata", "tsv_upload.yml.liquid", package = "RTD")
   m <- mock(0, 0)
   with_mock(
     exist_database = mock(FALSE),
@@ -18,17 +18,16 @@ test_that("td_upload works with mock",{
     tempdir = mock("/tmp"),
     `readr::write_tsv` = mock(TRUE),
     `Sys.which` = mock("/home/RTD/bin/embulk"),
-    `system2` = m,
-    {
-      td_upload(con, "test", "iris", iris)
+    `system2` = m, {
+      td_upload(con, "test", "iris", iris, mode = "embulk")
     }
   )
   expect_args(m, 1, embulk_exec, paste("guess", template_path, "-o /tmp/load.yml"))
   expect_args(m, 2, embulk_exec, "run /tmp/load.yml")
 })
 
-test_that("td_upload works with mock when the table already exists",{
-  template_path <- system.file("extdata", "tsv_upload.yml.liquid", package="RTD")
+test_that("td_upload works with mock when the table already exists", {
+  template_path <- system.file("extdata", "tsv_upload.yml.liquid", package = "RTD")
   m <- mock(0, 0)
   with_mock(
     exist_database = mock(FALSE, cycle = TRUE),
@@ -39,13 +38,11 @@ test_that("td_upload works with mock when the table already exists",{
     tempdir = mock("/tmp", cycle = TRUE),
     `readr::write_tsv` = mock(TRUE, cycle = TRUE),
     `Sys.which` = mock("/home/RTD/bin/embulk", cycle = TRUE),
-    `system2` = m,
-    {
+    `system2` = m, {
       expect_error(td_upload(con, "test", "iris", iris), ".* already exists.")
-      td_upload(con, "test", "iris", iris, overwrite = TRUE)
+      td_upload(con, "test", "iris", iris, mode = "embulk", overwrite = TRUE)
     }
   )
   expect_args(m, 1, embulk_exec, paste("guess", template_path, "-o /tmp/load.yml"))
   expect_args(m, 2, embulk_exec, "run /tmp/load.yml")
 })
-