+++ NEWS.md
@@ -1,3 +1,27 @@
+# RTD 0.4.1
+
+* Resubmit for CRAN. Only document update.
+
+# RTD 0.4.0
+
+* Replace msgpack-r to RcppMsgPack
+
+# RTD 0.3.0
+
+* Support dplyr 1.0.0
+
+# RTD 0.2.0
+
+* Fixed API key handling (#8)
+* Add bulk_import option which enables to upload dataframe without embulk (#11, #12, #14)
+* Add update_schema in table.R (#13)
+* Add append mode for embulk import (#9)
+* Stop releasing on CRAN
+
+# RTD 0.1.1
+
+* Add http_proxy configuration
+
 # RTD 0.1.0
 
 * Use embulk instead of TD toolbelt.